chore: move directory path
