+++ src/main/java/org/swmaestro/repl/gifthub/auth/entity/BaseTimeEntity.java
