+++ src/main/java/org/swmaestro/repl/gifthub/security/JpaUserDetailsService.java
@@ -1,11 +1,10 @@
-package org.swmaestro.repl.gifthub.auth.service;
+package org.swmaestro.repl.gifthub.security;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import org.swmaestro.repl.gifthub.auth.entity.CustomUserDetails;
 import org.swmaestro.repl.gifthub.auth.entity.Member;
 import org.swmaestro.repl.gifthub.auth.repository.MemberRepository;
